Make clear naming around subscribers, add notification update data structure
