+++ data_structures/point.go
@@ -1,0 +1,6 @@
+package data_structures
+
+type Point struct {
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
+}
