+++ data_structures/server_notification_subscriber.go
@@ -1,6 +1,6 @@
 package data_structures
 
-type Subscriber struct {
+type ServerNotificationSubscriber struct {
 	ID  string `json:"id" bson:"id"`
 	URL string `json:"url" bson:"url"`
 }
