+++ data_structures/server_notification_update.go
@@ -1,0 +1,6 @@
+package data_structures
+
+type NotificationUpdate struct {
+	Alert      Feature           `json:"alert,omitempty"`
+	SPCOutlook SPCOutlookProduct `json:"spcoutlook,omitempty"`
+}
